vista de reportes administrador modificados
